+++ sdk/resourcemanager/compute/armcompute/snapshots_client.go
@@ -48,7 +48,7 @@ func NewSnapshotsClient(subscriptionID string, credential azcore.TokenCredential
 // BeginCreateOrUpdate - Creates or updates a snapshot.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2022-07-02
+// Generated from API version 2023-01-02
 //   - resourceGroupName - The name of the resource group.
 //   - snapshotName - The name of the snapshot that is being created. The name can't be changed after the snapshot is created.
 //     Supported characters for the name are a-z, A-Z, 0-9, _ and -. The max name length is 80
@@ -71,7 +71,7 @@ func (client *SnapshotsClient) BeginCreateOrUpdate(ctx context.Context, resource
 // CreateOrUpdate - Creates or updates a snapshot.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2022-07-02
+// Generated from API version 2023-01-02
 func (client *SnapshotsClient) createOrUpdate(ctx context.Context, resourceGroupName string, snapshotName string, snapshot Snapshot, options *SnapshotsClientBeginCreateOrUpdateOptions) (*http.Response, error) {
 	req, err := client.createOrUpdateCreateRequest(ctx, resourceGroupName, snapshotName, snapshot, options)
 	if err != nil {
@@ -107,7 +107,7 @@ func (client *SnapshotsClient) createOrUpdateCreateRequest(ctx context.Context, 
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2022-07-02")
+	reqQP.Set("api-version", "2023-01-02")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, runtime.MarshalAsJSON(req, snapshot)
@@ -116,7 +116,7 @@ func (client *SnapshotsClient) createOrUpdateCreateRequest(ctx context.Context, 
 // BeginDelete - Deletes a snapshot.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2022-07-02
+// Generated from API version 2023-01-02
 //   - resourceGroupName - The name of the resource group.
 //   - snapshotName - The name of the snapshot that is being created. The name can't be changed after the snapshot is created.
 //     Supported characters for the name are a-z, A-Z, 0-9, _ and -. The max name length is 80
@@ -137,7 +137,7 @@ func (client *SnapshotsClient) BeginDelete(ctx context.Context, resourceGroupNam
 // Delete - Deletes a snapshot.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2022-07-02
+// Generated from API version 2023-01-02
 func (client *SnapshotsClient) deleteOperation(ctx context.Context, resourceGroupName string, snapshotName string, options *SnapshotsClientBeginDeleteOptions) (*http.Response, error) {
 	req, err := client.deleteCreateRequest(ctx, resourceGroupName, snapshotName, options)
 	if err != nil {
@@ -173,7 +173,7 @@ func (client *SnapshotsClient) deleteCreateRequest(ctx context.Context, resource
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2022-07-02")
+	reqQP.Set("api-version", "2023-01-02")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	return req, nil
 }
@@ -181,7 +181,7 @@ func (client *SnapshotsClient) deleteCreateRequest(ctx context.Context, resource
 // Get - Gets information about a snapshot.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2022-07-02
+// Generated from API version 2023-01-02
 //   - resourceGroupName - The name of the resource group.
 //   - snapshotName - The name of the snapshot that is being created. The name can't be changed after the snapshot is created.
 //     Supported characters for the name are a-z, A-Z, 0-9, _ and -. The max name length is 80
@@ -222,7 +222,7 @@ func (client *SnapshotsClient) getCreateRequest(ctx context.Context, resourceGro
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2022-07-02")
+	reqQP.Set("api-version", "2023-01-02")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
@@ -240,7 +240,7 @@ func (client *SnapshotsClient) getHandleResponse(resp *http.Response) (Snapshots
 // BeginGrantAccess - Grants access to a snapshot.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2022-07-02
+// Generated from API version 2023-01-02
 //   - resourceGroupName - The name of the resource group.
 //   - snapshotName - The name of the snapshot that is being created. The name can't be changed after the snapshot is created.
 //     Supported characters for the name are a-z, A-Z, 0-9, _ and -. The max name length is 80
@@ -265,7 +265,7 @@ func (client *SnapshotsClient) BeginGrantAccess(ctx context.Context, resourceGro
 // GrantAccess - Grants access to a snapshot.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2022-07-02
+// Generated from API version 2023-01-02
 func (client *SnapshotsClient) grantAccess(ctx context.Context, resourceGroupName string, snapshotName string, grantAccessData GrantAccessData, options *SnapshotsClientBeginGrantAccessOptions) (*http.Response, error) {
 	req, err := client.grantAccessCreateRequest(ctx, resourceGroupName, snapshotName, grantAccessData, options)
 	if err != nil {
@@ -301,7 +301,7 @@ func (client *SnapshotsClient) grantAccessCreateRequest(ctx context.Context, res
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2022-07-02")
+	reqQP.Set("api-version", "2023-01-02")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, runtime.MarshalAsJSON(req, grantAccessData)
@@ -309,7 +309,7 @@ func (client *SnapshotsClient) grantAccessCreateRequest(ctx context.Context, res
 
 // NewListPager - Lists snapshots under a subscription.
 //
-// Generated from API version 2022-07-02
+// Generated from API version 2023-01-02
 //   - options - SnapshotsClientListOptions contains the optional parameters for the SnapshotsClient.NewListPager method.
 func (client *SnapshotsClient) NewListPager(options *SnapshotsClientListOptions) *runtime.Pager[SnapshotsClientListResponse] {
 	return runtime.NewPager(runtime.PagingHandler[SnapshotsClientListResponse]{
@@ -351,7 +351,7 @@ func (client *SnapshotsClient) listCreateRequest(ctx context.Context, options *S
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2022-07-02")
+	reqQP.Set("api-version", "2023-01-02")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
@@ -368,7 +368,7 @@ func (client *SnapshotsClient) listHandleResponse(resp *http.Response) (Snapshot
 
 // NewListByResourceGroupPager - Lists snapshots under a resource group.
 //
-// Generated from API version 2022-07-02
+// Generated from API version 2023-01-02
 //   - resourceGroupName - The name of the resource group.
 //   - options - SnapshotsClientListByResourceGroupOptions contains the optional parameters for the SnapshotsClient.NewListByResourceGroupPager
 //     method.
@@ -416,7 +416,7 @@ func (client *SnapshotsClient) listByResourceGroupCreateRequest(ctx context.Cont
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2022-07-02")
+	reqQP.Set("api-version", "2023-01-02")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
@@ -434,7 +434,7 @@ func (client *SnapshotsClient) listByResourceGroupHandleResponse(resp *http.Resp
 // BeginRevokeAccess - Revokes access to a snapshot.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2022-07-02
+// Generated from API version 2023-01-02
 //   - resourceGroupName - The name of the resource group.
 //   - snapshotName - The name of the snapshot that is being created. The name can't be changed after the snapshot is created.
 //     Supported characters for the name are a-z, A-Z, 0-9, _ and -. The max name length is 80
@@ -458,7 +458,7 @@ func (client *SnapshotsClient) BeginRevokeAccess(ctx context.Context, resourceGr
 // RevokeAccess - Revokes access to a snapshot.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2022-07-02
+// Generated from API version 2023-01-02
 func (client *SnapshotsClient) revokeAccess(ctx context.Context, resourceGroupName string, snapshotName string, options *SnapshotsClientBeginRevokeAccessOptions) (*http.Response, error) {
 	req, err := client.revokeAccessCreateRequest(ctx, resourceGroupName, snapshotName, options)
 	if err != nil {
@@ -494,7 +494,7 @@ func (client *SnapshotsClient) revokeAccessCreateRequest(ctx context.Context, re
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2022-07-02")
+	reqQP.Set("api-version", "2023-01-02")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	return req, nil
 }
@@ -502,7 +502,7 @@ func (client *SnapshotsClient) revokeAccessCreateRequest(ctx context.Context, re
 // BeginUpdate - Updates (patches) a snapshot.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2022-07-02
+// Generated from API version 2023-01-02
 //   - resourceGroupName - The name of the resource group.
 //   - snapshotName - The name of the snapshot that is being created. The name can't be changed after the snapshot is created.
 //     Supported characters for the name are a-z, A-Z, 0-9, _ and -. The max name length is 80
@@ -524,7 +524,7 @@ func (client *SnapshotsClient) BeginUpdate(ctx context.Context, resourceGroupNam
 // Update - Updates (patches) a snapshot.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 2022-07-02
+// Generated from API version 2023-01-02
 func (client *SnapshotsClient) update(ctx context.Context, resourceGroupName string, snapshotName string, snapshot SnapshotUpdate, options *SnapshotsClientBeginUpdateOptions) (*http.Response, error) {
 	req, err := client.updateCreateRequest(ctx, resourceGroupName, snapshotName, snapshot, options)
 	if err != nil {
@@ -560,7 +560,7 @@ func (client *SnapshotsClient) updateCreateRequest(ctx context.Context, resource
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2022-07-02")
+	reqQP.Set("api-version", "2023-01-02")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, runtime.MarshalAsJSON(req, snapshot)